requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
oracledb==1.3.2
sqlalchemy==2.0.20
python-dotenv==1.0.0
aiofiles==23.2.1
gunicorn==21.2.0
pydantic==2.3.0
email-validator==2.0.0.post2